project/requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
bcrypt==3.2.2
billiard==3.6.4.0
blinker==1.4
cachelib==0.9.0
celery==5.2.7
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Caching==2.0.2
Flask-Cors==3.0.10
Flask-Login==0.6.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security-Too==4.1.4
Flask-SQLAlchemy==2.5.1
Flask-SSE==1.0.0
Flask-WTF==1.0.1
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
pkg_resources==0.0.0
prompt-toolkit==3.0.30
pycparser==2.21
pyparsing==3.0.9
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.39
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
